Fixup WDSL First Service to support SOAP Client
Expectations of the WDSL First Service to be running needed to have the
`server.port` set to one that would not conflict with the client
running on 8080.

The HashMap for the orders while the service is running needs to
properly increment the order for the customer.

Last, ignore Desktop Services files, `.DS_Stone` created by Mac OS